+++ src/test/java/test/config/FakeConfig.java
@@ -1,7 +1,0 @@
-package test.config;
-
-import org.springframework.context.annotation.Configuration;
-
-@Configuration public class FakeConfig {
-
-}