+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.3.0
-- Splitted REMREM Publish component into cli, service and common.
+- Split REMREM Publish component into cli, service and common.
 
 ## 0.2.9
 - Updated the RemRem shared version to latest.